pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Media Downloader with Telegram Bot"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.124.4",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.21",
    "uvicorn>=0.38.0",
    "yt-dlp>=2025.12.8",
    "python-telegram-bot>=20.8",
    "sqlalchemy>=2.0.25",
    "aiosqlite>=0.19.0",
    "jinja2>=3.1.3",
]
